backend/requirements_light.txt
# Lightweight requirements for 512MB Render free tier
setuptools>=65.0.0
wheel
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.12
requests==2.31.0
webcolors==1.13
Pillow==10.1.0
numpy==1.24.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.1

# Remove these heavy packages:
# tensorflow (200MB+) 
# opencv-python (100MB+)
# pandas (50MB+)
# scikit-learn (50MB+)
# keras, mediapipe, etc.
